Move the distributor implementation up a level (#612)
These need to be accessible from the omniwitness in order to distribute from there. I've also introduced an interface for the Witness instead of directly relying on the http client.
